requirements.txt
flask
requests
gunicorn
pytesseract
pillow
